Refactor calendar_attendance_page.dart. - Configure dropdown filter to be reusable and put into its own file. (drop_down_filter.dart). - Extract attendance header into its own file, which includes a drop down filter. - Extract calendar into own file to be reusable. - Create attendance_filter_enum.dart for filter options for DropdownFilter.
