+++ lib/enums/filter_base_enum.dart
@@ -1,0 +1,7 @@
+import 'package:flutter/material.dart';
+
+abstract class BaseEnum {
+    Color get color;
+    Color get borderColor;
+    String get label;
+}
